requirements.txt
sqlalchemy>=2.0
pydantic>=2.0
requests>=2.31
python-dotenv>=1.0
schedule>=1.2
pytest>=7.0
google-genai>=1.0
